Make Vecs in api response more ergonomic
In order to ensure a simpler api, this commit updates all Option<Vec>
fields to instead be Vec. This allows us to just have an empty Vec when
the response doesn't include any data. Because of this change, it's one
less level of option we have to deal with and it lends itself to a more
ergonomic api. It's easier to think of a list always existing even if
it's empty vs a list that may exist or not.